HttpSecurity.authorizeExchange() allows Method Chaining
Fixes gh-4397
